pyproject.toml
[project]
name = "amazon-partner-bot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "python-telegram-bot>=22.5",
    "requests>=2.32.5",
    "tenacity>=9.1.2",
]
